Split up the initialization of the PML
There are two parts to seting up the PML: figuring out where it is and
setting its material parameters. Previously those were both done
back-to-back and before `uservp` was called. However, sometimes a user
might need to know where the PML is when setting material parameters
(e.g. when determining the location of a sheet of graphene). This
splits the PML setup into two parts and calls `uservp` in the middle
to make this use-case possible.