Generalize the test script so that it works with arbitrary directories
The script now walks the file tree below where it is invoked, searches
for files named `tests.json`, and runs the tests described in any
files it finds. The current method of calling it from the top-level
NekCEM directory doesn't change.

This way we can use the same test framework to run tests in
`NekCEM-Examples`.