Add first acoustic solver test
It is adapted from the old `z_flat.usr` file in the SVN repo, and it
tests a flat interface with PEC boundary conditions. The code needs
more cleaning up, but I'm trying to initially commit the file with the
minimal changes needed to make it run.

This also fixes a bug introduced into the acoustic solver in
856870ee4948490ece3fac1d2b6e74803cb59dfc where a complex square root
was incorrectly changed to a norm.