Give `stylecheck` the option to also try and fix issues
If passed the option `--fixup`, the style-checking script will now try
to fix any issues it finds. The original file will be overwritten with
a corrected version and the old version will be backed up as
`<file>.old`. Note that the script should be used with caution: if a
line exceeds the allowed Fortran character length due to respacing the
script will not do anything to fix it.