Allow for a proper installation of NekCEM
Installation is done with `make install`. Our install is a little
abnormal since there is no code to compile, but there is still value
in copying `NekCEM/src` and `NekCEM/bin` to `/usr/local` and then
symlinking `configurenek` and `nek` to `/usr/local/bin` because then
`configurenek` and `nek` are on the user's PATH. Then on a typical
system the build process is just

- configurenek
- make
- mpiexec

For developers also allow for an "in place build" which just symlinks
`configurenek` and `nek` to the in-tree executables.